Fixed layout issue
