+++ frontend/app/layout.tsx
@@ -1,0 +1,41 @@
+import MobileNavbar from "@/components/mobile/MobileNavbar";
+import Navbar from "@/components/navbar/Navbar";
+import UserContext from "@/components/utility/UserContext";
+import GrafanaFaro from "@/observability/GrafanaFaro";
+import type { Metadata } from "next";
+import { Inter, Roboto_Condensed } from "next/font/google";
+import "./globals.css";
+
+const inter = Inter({ subsets: ["latin"] });
+
+export const metadata: Metadata = {
+  title: "Sports Hub",
+  description: "Generated by create next app",
+};
+
+const roboto_condensed = Roboto_Condensed({
+  weight: "400",
+  subsets: ["latin"],
+  display: "swap",
+  variable: "--font-roboto-condensed",
+});
+
+export default function RootLayout({ children }: { children: React.ReactNode }) {
+  return (
+    <html lang="en">
+      <GrafanaFaro>
+        <UserContext>
+          <body className={`${inter.className} ${roboto_condensed.variable}`}>
+            <div className="hidden md:block">
+              <Navbar />
+            </div>
+            <div className="md:hidden">
+              <MobileNavbar />
+            </div>
+            {children}
+          </body>
+        </UserContext>
+      </GrafanaFaro>
+    </html>
+  );
+}
